requirements.txt
six==1.10.0
cmd2==1.0.1
python-tvmaze==1.0.1
ratelimit==2.2.1
tinydb==3.15.2
terminaltables==3.1.0
python-dateutil==2.8.1
typing-extensions==3.7.4.2
